Alternative to global server require path, add a preload.rb
This adds a preload (bootstrap) to give the bin stub just enough to
configure itself before the app loads and runs the action subscriber
railtie and load hooks. This solution should cover edge cases around
rails eager loading causing subscribers to load before load hooks have
been run under server mode.
